requirements.txt
# Core HTTP + DB
requests>=2.28.0
pymongo>=4.16.0
dnspython>=2.6.1
python-dotenv>=1.0.0

# Web app + production server (Ubuntu VM)
Flask>=3.0.0
gunicorn>=21.2.0

# Gemini + Google GenAI
google-generativeai>=0.3.0
google-genai>=1.55.0

# Geo/timezone & user-agent (for future reuse of DEBot-style enrichment)
geopy>=2.4.1
timezonefinder>=8.1.0
fake-useragent>=2.2.0

# CLI UI enhancements
colorama>=0.4.6

# Polymarket CLOB API client
py-clob-client>=0.1.0
